Reworked the solution to include Core-Web
Resworked the solution to include core-web project as well as updated it
from 1.0.1 to the latest 1.1.0's. The inclusion of the web project
solved problems with the native dependency injection and made it easier
to troubleshoot, so it will stay in here for a while. Posting commands
now works from rest to sql, so this is a nice milestone.